refactor(storage): extract identities module (R.2)
Phase R.2 of the storage modularization plan
(SPEC_STORE_MODULARIZATION_2026_05_27.md). Largest extraction so
far — moves the entire identity subsystem out of store.rs into a
new `identities.rs` (485 LOC):

Structs:
- `SecretRef` (enum)
- `IdentityAccount` + `default_context_json` / `default_identity_status`
- `Identity` (the named bundle)
- `IdentityBinding` (bundle ↔ account junction)
- `AgentIdentityLink` (agent ↔ account junction, legacy)

Methods (14 total across two impl Store{} blocks now consolidated):
- `identity_list` / `identity_get` / `identity_upsert` / `identity_delete`
- `agent_identity_link` / `agent_identity_unlink` /
  `agent_identity_list_for_agent`
- `bundle_identity_list` / `bundle_identity_get` /
  `bundle_identity_upsert` / `bundle_identity_delete`
- `bundle_identity_bind` / `bundle_identity_unbind` /
  `bundle_identity_bindings`

Method surface is unchanged: `Store::identity_*`,
`agent_identity_*`, `bundle_identity_*` still live on `Store` via
the new module's `impl Store {}` block. All 5 structs are
re-exported from `storage::store::*` AND `storage::*` so existing
imports stay green (3 external callers verified:
`identity/migration.rs`, `identity/resolver.rs`,
`server/identity_handlers.rs`).

Net store.rs change: -443 lines (6042 → 5599). New file 485 LOC.

162/162 storage tests + 51/51 identity tests green.
