+++ .changesets/1779973768-refactor-storage-extract-identities-module-r-2-86pe.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(storage): extract identities module (R.2)
